Update simulations.py
change code for get_plot_title.

Replace the colon in each title value with a blank space. The reason is that the title would like to be used as a folder name. In Windows, the name of folder is not allowed contains a semicolon, so there is always some errors.